[implode-decoder] Fix improper repeated literal writing.
In cases where repeated literals overlapped, the previous code's usage
of Buffer#copy was improper. Buffer#copy is advertised as "working with
overlapping reads/writes on the same buffer", but what it means by this
is "will copy the data that was previously there over, without being
interfered with by the intervening write". This is not what we want for
implode, rather, we want to always pull the latest byte value when its
written.